pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdfmill"
version = "0.1.0"
description = "Configurable PDF processing pipeline for splitting, transforming, and printing PDFs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yifei" }
]
keywords = ["pdf", "printing", "labels", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Printing",
]
dependencies = [
    "pypdf>=4.0",
    "pyyaml>=6.0",
    "pywin32>=307; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
ocr = [
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]

[project.scripts]
pdfp = "pdfmill.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pdfmill"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "-ra"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "windows: marks tests that only run on Windows",
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src/pdfmill"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
fail_under = 80